SECCIÓN 13 - TERMINADO

+++ 03-admin-todos/src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserServerSession } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse } from 'next/server';
 import * as yup from 'yup';
@@ -31,9 +32,12 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) { 
 
+  const user = await getUserServerSession();
+  if( !user ) return NextResponse.json('No autorizado', { status: 401 });
+
   try {
     const { complete, description } = await postSchema.validate( await request.json() );
-    const todo = await prisma.todo.create({ data: { complete, description } });
+    const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } });
 
     return NextResponse.json( todo );
   } catch (error) {
@@ -43,8 +47,11 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE( request: Request ) { 
+  const user = await getUserServerSession();
+  if( !user ) return NextResponse.json('No autorizado', { status: 401 });
+
   try {
-    await prisma.todo.deleteMany({ where: { complete: true } });
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
     
     return Response.json({ message: 'Todos Completados Se Han Borrado', status: 200 });
   } catch (error) {
